% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{checkGlobalVars}
\alias{checkGlobalVars}
\title{Check for Use of Global Variables in a Function}
\usage{
checkGlobalVars(f, silent = FALSE)
}
\arguments{
\item{f}{The function to be checked for global variable usage.}

\item{silent}{Logical parameter with a default value of \code{FALSE}.
If \code{TRUE}, the function suppresses warnings about global variable usage.}
}
\value{
Returns \code{TRUE} if no global variables are used in the function \code{f}, \code{FALSE} otherwise.
If \code{silent} is \code{FALSE} and global variables are found, a warning is issued.
}
\description{
This function checks whether the specified function (\code{f}) uses any global variables.
It returns \code{TRUE} if no global variables are used, and \code{FALSE} otherwise. If global variables are found
and \code{silent} is \code{FALSE}, a warning is issued listing the global variables.
}
\examples{
testFunction <- function(x, y) { z <- x + y; return(z) }
checkGlobalVars(testFunction)
checkGlobalVars(testFunction, silent = TRUE)
}
