% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{checkVars}
\alias{checkVars}
\title{Check Variables in an Environment}
\usage{
checkVars(
  variables,
  envir,
  verbose = FALSE,
  prefix = "Problematic variables!\\n",
  suffix = NULL
)
}
\arguments{
\item{variables}{A character vector of variable names to check.}

\item{envir}{The environment in which to look for the variables.}

\item{verbose}{Report on good (passing) variables.}

\item{suffix}{Suffix to append to printed summary statment.}
}
\value{
No return value, called for side effects.
}
\description{
This function iterates over a list of variable names and checks their
existence and value in a given environment. It issues warnings for variables
that are missing, NULL, NA, NaN, infinite, or empty, and sends a message for
variables that are defined and not empty.
}
\examples{
myEnv <- new.env()
myEnv$aaa <- 111
myEnv$xxx <- NULL
myEnv$yyy <- list()
myEnv$zzz <- numeric()
output.variables <- c("aaa", "xxx", "zzz", "yyy", "bbb")
checkVars(output.variables, envir = myEnv)
}
