% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{.filterFunctionsFromObjNames}
\alias{.filterFunctionsFromObjNames}
\title{Check Names for Variable or Function Type}
\usage{
.filterFunctionsFromObjNames(names, envir)
}
\arguments{
\item{names}{A character vector containing names to check.}

\item{envir}{The environment in which to check for names.}
}
\value{
A character vector of variable names.
}
\description{
This function takes a character vector of object names and checks whether
they correspond to variables or functions within the provided environment.
It issues warnings for function names and for missing objects, and returns a list of variable names.
}
\examples{
myEnv <- new.env()
myEnv$x <- 4
myEnv$fff <- function(x) x^3
.filterFunctionsFromObjNames(c('x', 'fff'), envir = myEnv)
}
