% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{.importPackageFunctions}
\alias{.importPackageFunctions}
\title{Import all exported functions from a package into an environment}
\usage{
.importPackageFunctions(pkg, env)
}
\arguments{
\item{pkg}{A character string specifying the name of the package to import functions from.
Default: none.}

\item{env}{The environment where the functions should be imported. Default: none.}
}
\value{
This function does not return a value but imports the exported objects from the package
into the specified environment.
}
\description{
The \code{.importPackageFunctions()} function imports all exported objects from a specified package
into a given environment. This can be useful when you want to have direct access to all functions
of a package without explicitly calling the package name in each call.
}
