% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.other.R
\name{removeAllExceptFunctions}
\alias{removeAllExceptFunctions}
\title{Remove Non-function Objects}
\usage{
removeAllExceptFunctions(envir = .GlobalEnv)
}
\arguments{
\item{env}{environment from which to remove the variables, defaulting to the global environment.}
}
\value{
Nothing, but modifies the environment in place.
}
\description{
Removes all objects that are not functions from the specified environment.
}
\examples{
# Define some variables and functions in the global environment
x <- 1
y <- "a"
z <- list(1, 2, 3)
f <- function(x) x + 1
g <- function(y) y * 2

# Check the names of the objects in the global environment
ls()
# [1] "f" "g" "x" "y" "z"

# Remove all objects except functions from the global environment
removeAllExceptFunctions()

# Check the names of the objects in the global environment again
ls()
# [1] "f" "g"
}
