% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{sourceScriptAndPassFunctions}
\alias{sourceScriptAndPassFunctions}
\title{Source Script in a New Environment and Pass Functions}
\usage{
sourceScriptAndPassFunctions(scriptPath)
}
\arguments{
\item{scriptPath}{The path to the R script that should be sourced.}
}
\value{
A new environment containing the sourced script and copied functions.
}
\description{
This function creates a new environment with the base environment as its parent,
sources an R script into this new environment, and then copies all functions
from the global environment to the new environment.
}
\examples{
\dontrun{
  myEnv <- sourceScriptAndPassFunctions('~/path/to/script.R')
}
}
