% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{strict}
\alias{strict}
\title{Enforce Strict Evaluation in a Function}
\source{
From: moodymudskipper at https://stackoverflow.com/questions/6216968/r-force-local-scope/
}
\usage{
strict(f1, ...)
}
\arguments{
\item{f1}{The function for which strict evaluation is to be enforced.}

\item{...}{Additional arguments to be passed to the modified function.}
}
\value{
The result of executing the modified function with strict evaluation.
}
\description{
Modifies an existing function (\code{f1}) to enforce strict evaluation of its arguments.
This is achieved by redefining the function with its arguments evaluated in a specified environment,
which can help in debugging and ensuring that all variables are defined before function execution.
}
\examples{
# Define a sample function
Z <- 2
funOK <- function(x, y) { x + y }
funBAD <- function(x, y) { x + Z }
funOKwoParenthesis <- function(x, y) x + 1

# Check strict evaluation
strict(funOK, 1, 2)
strict(funBAD, 1, 2)
strict(funOKwoParenthesis, 1, 2)
}
