% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoENV.R
\name{.findFunctions}
\alias{.findFunctions}
\title{Find Functions in Specified Packages}
\usage{
.findFunctions(
  std_packages = c("ggplot"),
  custom_packages = c("Stringendo", "ReadWriter", "CodeAndRoll2", "MarkdownHelpers",
    "MarkdownReports", "Seurat.utils", "isoENV", "UVI.tools", "Connectome.tools",
    "NestedMultiplexer"),
  other_packages = NULL
)
}
\arguments{
\item{packages}{A character vector of package names.}
}
\value{
A list where each element is a character vector of function names for the corresponding package.
Packages not loaded or non-existent are returned as \code{NULL}.
}
\description{
This function returns a list of all functions available in the specified packages.
If a package is not loaded or does not exist, it is skipped.
}
\examples{
# Assuming the required packages are installed and loaded
pkgs <- c("ggplot2", "dplyr")
.findFunctions(pkgs)
}
